pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-api-portal"
version = "0.1.0"
authors = [
  { name="Developer", email="developer@example.com" },
]
description = "A modern API developer portal for Django REST Framework projects."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "django>=3.2",
    "djangorestframework>=3.12",
    "drf-spectacular>=0.26"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Framework :: Django",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/example/modern-drf-swagger"
"Bug Tracker" = "https://github.com/example/modern-drf-swagger/issues"
